added most/least side labels and fixed ordering
Just remembered the bit of feedback from class about the ordering of beers being unclear.
I added "most(/least) bitter(/sweet)" text that will appear on the side when the user presses either the "Bitterness" or "Sweetness" sort button. 
In addition, I changed the ordering of beer to descending instead of ascending so that the MOST sweet/bitter beer appears at the top of the list as opposed to the bottom. Oops!

Since this is a last-minute change, I commented the changes with "// ALENA EDIT", so you can easily remove the new code and revert to old code if this does not work/is not something we want.